requirements.txt

## keras ##
#tensorflow>=1.4.0
tqdm
colorama
regex
keras
gensim
h5py

## both ##
gtts
google-api-python-client
google-cloud-speech
PyAudio ## install this from linux distro repository, not pip3
unidecode
matplotlib
py-cpuinfo
datetime

## t2t ##
Tensor2Tensor
tensorflow-hub

## bert ##
bert-serving-server
bert-serving-client
numpy

## kaggle ##
kaggle
## must install kaggle.json file as described here:
## https://github.com/Kaggle/kaggle-api

## pytorch ##
## go to the pytorch web site to download the latest version of pytorch ##

## text-to-speech ##
## install mpg321 from your linux repository ##
## install python3-pyaudio
## install python-pyaudio

## tensorflow-model-server ##
## see this url:
## https://www.tensorflow.org/tfx/serving/setup#installing_the_modelserver
## this python package is also required
tensorflow-serving-api